swap xml formatter, refactor
