+++ package.json
@@ -33,6 +33,6 @@
     "chalk": "^2.3.0",
     "commander": "^2.12.2",
     "express": "^4.16.2",
-    "xml2json": "^0.11.2"
+    "xml2js": "^0.4.23"
   }
 }
